Asking for more flexibility in predict function
Hello, 

Thanks very much for accepting my previous pull request "Update shap.py #49" about giving a choice to the user in whether to use link = "identity" or link = "logit".

I have another proposal. On line 197, predict_func is set to model.predict_proba if the model is a classifier and is a BaseEstimator. I believe, that just as with the link, users should be given a choice about whether to use model.predict_proba or model.predict regardless of whether the model is a classifier or a regressor. This is because if predict == False, there is a runtime error for classification models like SVC.

Please let me know if this is a misunderstanding on my part or if there is a good reason why this was not done.

On another note, thank you for guiding me through my first pull request. 
 
Sincerely,
Shreyan